+++ jschon/exceptions.py
@@ -1,26 +1,2 @@
-class CatalogError(Exception):
-    pass
-
-
-class JSONError(Exception):
-    pass
-
-
-class JSONPatchError(Exception):
-    pass
-
-
-class JSONPointerError(Exception):
-    pass
-
-
-class JSONSchemaError(Exception):
-    pass
-
-
-class RelativeJSONPointerError(Exception):
-    pass
-
-
-class URIError(Exception):
-    pass
+# noinspection PyUnresolvedReferences
+from .exc import *