+++ jschon/jsonpointer.py
@@ -5,7 +5,12 @@ import re
 import urllib.parse
 from typing import Any, Iterable, Literal, Mapping, Sequence, TYPE_CHECKING, Union, overload
 
-from jschon.exceptions import JSONPointerError, RelativeJSONPointerError
+from jschon.exc import (
+    JSONPointerMalformedError,
+    JSONPointerReferenceError,
+    RelativeJSONPointerMalformedError,
+    RelativeJSONPointerReferenceError,
+)
 
 if TYPE_CHECKING:
     from jschon.json import JSON
@@ -74,8 +79,8 @@ class JSONPointer(Sequence[str]):
 
         :param values: each value may either be an RFC 6901 string, or an iterable
             of unescaped keys
-        :raise JSONPointerError: if a string argument does not conform to the RFC
-            6901 syntax
+        :raise JSONPointerMalformedError: if a string argument does not conform to
+            the RFC 6901 syntax
         """
         self = object.__new__(cls)
         self._keys = []
@@ -83,7 +88,7 @@ class JSONPointer(Sequence[str]):
         for value in values:
             if isinstance(value, str):
                 if not JSONPointer._json_pointer_re.fullmatch(value):
-                    raise JSONPointerError(f"'{value}' is not a valid JSON pointer")
+                    raise JSONPointerMalformedError(f"'{value}' is not a valid JSON pointer")
                 self._keys.extend(self.unescape(token) for token in value.split('/')[1:])
 
             elif isinstance(value, JSONPointer):
@@ -179,8 +184,8 @@ class JSONPointer(Sequence[str]):
         will always fail.
 
         :param document: any Python object
-        :raise JSONPointerError: if `self` references a non-existent location
-            within `document`
+        :raise JSONPointerReferenceError: if `self` references a non-existent
+            location in `document`
         """
 
         def resolve(value, keys):
@@ -203,7 +208,7 @@ class JSONPointer(Sequence[str]):
             except (KeyError, IndexError):
                 pass
 
-            raise JSONPointerError(f"Path '{self}' not found in document")
+            raise JSONPointerReferenceError(f"Path '{self}' not found in document")
 
         return resolve(document, collections.deque(self._keys))
 
@@ -283,13 +288,13 @@ class RelativeJSONPointer:
             a value of 0, which is not allowed by the grammar, is treated as if
             there is no adjustment.
         :param ref: a :class:`JSONPointer` instance, or the literal ``'#'``
-        :raise RelativeJSONPointerError: for any invalid arguments
+        :raise RelativeJSONPointerMalformedError: for any invalid arguments
         """
         self = object.__new__(cls)
 
         if value is not None:
             if not (match := RelativeJSONPointer._regex.fullmatch(value)):
-                raise RelativeJSONPointerError(f"'{value}' is not a valid relative JSON pointer")
+                raise RelativeJSONPointerMalformedError(f"'{value}' is not a valid relative JSON pointer")
 
             up, over, ref = match.group('up', 'over', 'ref')
             self.up = int(up)
@@ -337,25 +342,25 @@ class RelativeJSONPointer:
         node = document
         for _ in range(self.up):
             if node.parent is None:
-                raise RelativeJSONPointerError('Up too many levels')
+                raise RelativeJSONPointerReferenceError('Up too many levels')
             node = node.parent
 
         if self.over:
             if node.parent is None:
-                raise RelativeJSONPointerError('No containing node for index adjustment')
+                raise RelativeJSONPointerReferenceError('No containing node for index adjustment')
             if node.parent.type != "array":
-                raise RelativeJSONPointerError(f'Index adjustment not valid for type {node.parent.type}')
+                raise RelativeJSONPointerReferenceError(f'Index adjustment not valid for type {node.parent.type}')
             adjusted = int(node.key) + self.over
             if adjusted < 0 or adjusted >= len(node.parent):
-                raise RelativeJSONPointerError(f'Index adjustment out of range')
+                raise RelativeJSONPointerReferenceError(f'Index adjustment out of range')
             node = node.parent[adjusted]
 
         if self.index:
             if node.parent is None:
-                raise RelativeJSONPointerError('No containing node')
+                raise RelativeJSONPointerReferenceError('No containing node')
             return int(node.key) if node.parent.type == "array" else node.key
 
         try:
             return self.path.evaluate(node)
-        except JSONPointerError as e:
-            raise RelativeJSONPointerError from e
+        except JSONPointerReferenceError as e:
+            raise RelativeJSONPointerReferenceError from e